Hardcoded and Fixed Grocery Page
Hardcoded the names and pictures of each grocery item, added the button
dropdown, and fixed the pictures in the image file to all the same
format: lower case and spaces where needed.  Also added the picture
locations to the ingredients excel file for easy database entry.
Created two sets of buttons to see which one works better.  Could not
find a way to make the two red buttons expand the width of their
thumbnail container.